move test.db to examples folder, move main_db.py into main.py
